Hide "Email link" and "Phone link" from selections in CMS
